+++ CHANGES.md
@@ -7,6 +7,7 @@ wavesurfer.js changelog
 - add `hideCursor` option to hide the mouse cursor when hovering over the waveform (#2367)
 - Regions plugin: check `maxLength` before resizing region (#2374)
 - Regions plugin: Add support for drag selection to be separated for each channel (#2380)
+- add optional `channelIdx` parameter to `setWaveColor`, `getWaveColor`, `setProgressColor`, `getProgressColor` methods (#2391)
 
 5.2.0 (16.08.2021)
 ------------------